+++ index.html
@@ -230,6 +230,6 @@
     </div>
   </template>
 
-  <script src="js/main.js" type="module" defer></script>
+  <script src="js/main.js" type="module"></script>
 </body>
 </html>
