--- js/choose.js
@@ -1,1 +1,0 @@
-// for choosing offer on a map
