--- js/comparison.js
@@ -1,1 +1,0 @@
-// for offer comparison
