--- js/form-submit.js
@@ -1,1 +1,0 @@
-// for form submiting anf getting offers
