+++ js/main.js
@@ -1,7 +1,3 @@
-import {getOffer} from './get-offer.js';
-import {NUMBER_OF_OFFERS} from './data.js';
-
-const offers = new Array(NUMBER_OF_OFFERS).fill(null).map(getOffer);
+import {offers} from './data.js';
 
 offers;
-console.log('aa');
