--- js/page-state.js
@@ -1,1 +1,0 @@
-// for setting page state - active||none
