--- js/validation.js
@@ -1,1 +1,0 @@
-// for form validation
